Add EGL backend for ruby
Unlike GLX, EGL is not tied to X. This may be necessary to support OpenGL on Wayland in the future.
Unfortunately, it is not possible to build both the EGL backend and the GLX backend at the same time because ruby/video/opengl/*.hpp headers expect glGetProcAddress (either the GLX or the EGL version) as a compile-time macro.